requirements.txt
tabulate
request